rt: update check in scheduleChooseNewThread
It was decided that for the MCS kernel, the appropriate check is
not ksDomainTime == 0 but isCurDomainExpired because time has not yet
been committed and ksConsumed may be non-zero. It was also decided
that it doesn't make sense for MIN_BUDGET to be used with domains
and so it was removed from isCurDomainExpired.

Signed-off-by: Mitchell Buckley <Mitchell.Buckley@data61.csiro.au>
Co-authored-by: Curtis Millar <2741168+xurtis@users.noreply.github.com>